HFP-1653 Add (preliminary) function to get common fields from semantics
Semantics of main content and subcontent can be parsed for "common" fields
(or other field combinations, since the function is general).
The idea was to not touch the core, but this way all libraries will have
to be loaded and it might take 3-4 seconds until the interface for
all translations is fully populated.

It's propably better to find a way to grab the required information by
modifying h5peditor.js in core.