HFP-1577 Add temporary workaround for crash in canvas
When the canvas has loaded and nodes were added, this would lead
to a crash if the translations had not loaded already (since working
on the same data). Enabling the canvas is now delayed until the
translations have been loaded.

The whole library/semantics/translation loading across components
should be simplified and made faster.