HFP-2295 Change main selector for linked bq to always be old-content
Previously new content would always show the blank option for branching
question main selector. Now it will show "jump to another branch" when
the content is linked, even if this is new content, which makes more
sense. New content may be initially linked if it is inserted before
otter content.