HFP-2330 Change behavior when deleting node that was branched to
When the user deleted a node (a) that some other node (b) was
branching to, (b) instead linked to the successor of (a) or to
the default end scenario if the successor was (b).
The new behavior is to branch to always branch to the default
end scenario if (b) is branching to (a) and a is deleted.