HFP-1574 Fix requiring optional dependencies for editor.
If the optional dependencies of branching scenario is not explicitly
declared we will not be able to render the preview because of the way
H5P Core requiring dependencies work. When we try to load a library
it only loads the dependencies of that library, we're assuming that
the libraries have already been resolved and added for the view by
core logic, but this is not true for the editor, thus we have to
explicitly declare the dependencies like this unfortunately.